feat(storybook-utils): use decorator for v-models instead of render function (#67)
closes #49 

Refactored `@sit-onyx/storybook/utils` to use a Storybook decorator
instead of a custom render function for implementing the v-models. This
has the following improvements:
- improved code snippet generation which removes unnecessary v-bind for
props and event handlers
- old:
![image](https://github.com/SchwarzIT/onyx/assets/67898185/e6931991-b908-417a-8536-f310dc689767)
- new:
![image](https://github.com/SchwarzIT/onyx/assets/67898185/0a97623e-0866-46b2-b5d1-75d241362905)
- the property value is now also reflected to the vue component itself
and not only the Storybook control table
- old:
![image](https://github.com/SchwarzIT/onyx/assets/67898185/4598d428-ca7a-4cd8-aafd-6def35c09424)
- new:
![image](https://github.com/SchwarzIT/onyx/assets/67898185/2b587cbc-1328-451d-8f32-3236a5ca0e7a)
- since we no longer use the render function, we no longer have
conflicts if we define a render function for a specific story in the
future to showcase component slots

### Non-primary stories

On the docs page, only the first (primary) story and the story below the
property table (which is also the same primary story as the first one)
will update the Storybook property table. All other stories on the same
page are isolated and will not affect the property table and/or other
stories. Imho this is the intended behavior.


![image](https://github.com/SchwarzIT/onyx/assets/67898185/02584858-1be5-483e-ade8-89ad37b79796)