chore: prevent workflows from overriding auto created PRs (#105)
## Description

Previously all Pull requests that are created via workflows (e.g.
Playwright screenshots, Figma variables) will override/merge each other.
This leads to the issue that if you e.g. trigger a PR to update
screenshots and afterwards one for the Figma variables, the result will
be 1 PR with both changes which is unintentional.