feat: link invalid to custom errorMessage (#151)
## Description

By linking the errorMessage prop to the setCustomValidity function of
the input element, the devs that use the component can trigger an
invalid state by their own validation logic whenever needed.
The validityChange event includes all information of invalid states, so
when the projects e.g. want to provide a specific error message on
patternMismatch, both patternMismatch and customError will turn true in
the ValidityState.