feat: create alpha-test-app (#157)
During the alpha phase of onyx, we can make use of a bare vitest starter
app to try out onyx components in a user-near environment and detect
problems that Storybook can not show.
The alpha-test-app will be removed once onyx leaves the alpha phase. As
it is temporary, it will not be mentioned in our documentation app.