feat(grid-layout): implemented class generation for grid layouting (#195)
Implements the grid layout as defined by UX.
We provide the following classes:

- `onyx-grid`: Sets up the Grid wrapper, which configures the grid
layout depending on the breakpoint. Should be set on the main element of
an application.

Additionally there are some optional modifier classes available. They
all must be set with and on the same element as the `onyx-grid` class.

- `onyx-grid-xl-20`: Increases the column count from 16 to 20 for the
`xl` breakpoint.
- `onyx-grid-max-md`:  Caps the width before the `lg` breakpoint
- `onyx-grid-max-lg`:  Caps the width before the `xl` breakpoint