docs: exclude utility/support components from sidebar (#282)
closes #284

Only show components inside the Sidebar, no utility or support
components.
Also add a link to the Storybook in the headers "Resources" dropdown so
the utility/support components can be seen there.

Fixed the bug that components with a folder other than "Components" are
not displayed correctly, e.g. `utility/GridPlayground`.