perf: use self hosted runner for Playwright tests (#661)
Relates to #644

I set up a self hosted GitHub runner in STACKIT which will be used for
the Playwright screenshot tests to improve performance.

The other workflows like check code quality, release etc. are still
using the default GitHub runner so we can use the full performance of
the STACKIT runner for the screenshots and it is not busied with
"simple" stuff that the default runner is perfectly fine for.

### Notes
Since the self hosted runner is only 1 VM and no kubernetes or
something, multiple Playwright workflows have to wait and can not be run
in parallel (should not happen too often I guess and even if they wait,
they are still faster than if they would run in parallel).