fix(OnyxTooltip): add missing type for open property (#775)
Relates to #567 

When using the `OnyxTooltip` in a project outside the onyx monorepo, the
type for the `open` property is just `any` because it was imported from
the `@sit-onyx/headless` package which is NOT a peerDependency of
`sit-onyx` so it could not be resolved. I copied the type and removed
the import (I didn't want to introduce the peerDependency just for a
single type).

I also excluded the `stories.d.ts` files from the build output (dist
folder) because they are not needed.