feat(OnyxSelect): initiate input area of OnyxSelect (#759)
Creates the base for OnyxSelect. This includes labels, message, and the
central input-like section which will trigger a flyout in the future.
For now, the model is based on strings for showcasing the selections,
this will be replaced by whatever types ListBox and ComboBox will need
(they are not finished yet).

Also, a minor style discrepency in OnyxInput has been fixed.