fix(OnyxIcon): revert default size back to 24px (#1406)
Relates to #1307

We changed the default icon size to match the parent font size in #1307
which caused a lot of unintended changes in Storybook examples and our
docs. Also to match the parent font size is a rare use case so we change
back the default size back to 24px to better fit common use cases.