feat(nuxt): integrate @nuxtjs/i18n into onyx module (#1402)
Closes #1345 

This PR adds an automatic integration for the @nuxtjs/i18n module to the
onyx nuxt module.
It allows vue-i18n to take over the translation work giving users access
to all its features like e.g. lazy loading of translations.
A dedicated section regarding this was added to the docs of the nuxt
module.