feat(OnyxSelect, OnyxRadioGroup, OnyxCheckboxGroup): Implement truncation on higher level (#1541)
<!-- Is your PR related to an issue? Then please link it via the
"Relates to #" below. Else, remove it. -->

Relates to #1539 

Up until now the truncation property was on option-level only for
OnyxSelect, OnyxRadioGroup and OnyxCheckboxGroup. The user had to set
the property for each option separately. In this PR, the truncation
property is added to the grouped component too, so that the user can set
the truncation only once for every option. This property is used only as
a fallback if the truncation is not set on option-level.

## Checklist

- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>