feat(OnyxSelect): Implement OnyxSelect open direction and alignment (#1745)
<!-- Is your PR related to an issue? Then please link it via the
"Relates to #" below. Else, remove it. -->

Relates to #1198 

Added new feature which is calculating the space above and below the
component and based on that the flyout will open to the top or bottom. I
also added new 'alignment' property which will align the flyout to the
right, left or will set the width to match the width of the input.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Lars Rickert <lars.rickert@mail.schwarz>