fix(OnyxMiniSearch): Adjust the size to the search input (#1777)
<!-- Is your PR related to an issue? Then please link it via the
"Relates to #" below. Else, remove it. -->

Relates to #1198 

The native input component has a default property size with value 20,
which is approximately 100px. This was preventing us to make the input
smaller. I adjusted the size property to 1 for the purposes of
OnyxMiniSearch component.

## Checklist

- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>