feat(icons): implement changeset generation (#1794)
Relates to #939

Generate changeset with list of new, deleted and modified icons whenever
we import icons from Figma so they are released as npm package.

I also removed UX from codeowners for the icons so they are not added as
required reviewer. The UX review process will be done in Figma so we
devs should be responsible to approve the technical import via CI to
make sure everything is imported correctly.