docs: fix charts not rendering because of version mismatch (#1823)
Closes #1822

- fix charts not rendering because of version mismatch
- **Cause:** This was caused by mismatching versions of the `chart.js`
package in our `docs` and the `@sit-onyx/chartjs-plugin` packages.
- **Fix:** We now make use of the [PNPM catalog
feature](https://pnpm.io/catalogs) to enforce usage of the same package
version. Added screenshot tests to ensure that we catch these issues in
the future.

- fix warning logged by vitepress about `gobalThis.document` not being
available
- **Cause:** Vitepress does server-side rendering, where the document is
not available. But we try to set a CSS variable.
- **Fix:** We use the optional chaining operator to ensure that we set
the CSS variable only when `gobalThis.document` is defined.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: JoCa96 <18483883+JoCa96@users.noreply.github.com>