perf: improve Storybook performance (#1853)
Update Storybook to version 8.3.0-alpha.5 which includes a bug fix which
reduced the bundle size of our build Storybook from 26MB to 6MB and
reduced build time from 37s to 10s.

This change also significantly improves the loading performance of the
onyx Storybook.

If you are interested, the bug was caused by the vue-component-meta
package which included ALL schema definitions for props, events, slots
and exposed so if e.g. HTMLElement was used, it included all nested
props, JSDocs etc. although they are not needed/used by Storybook, see
[Storybook PR](https://github.com/storybookjs/storybook/pull/28589)