fix(OnyxRadioGroup, OnyxCheckboxGroup): Fix skeleton prop injection for radio and checkbox group (#1990)
<!-- Is your PR related to an issue? Then please link it via the
"Relates to #" below. Else, remove it. -->

Relates to #1967 

OnyxRadioGroup and OnyxCheckboxGroup are using a number for the skeleton
prop, whereas the provided skeleton prop from the composable is a
boolean. I've added the number type to the composable and now when a
component needs a fallback number for this property the default value
will be 3.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
is added