chore: fix dev script of all packages to use turborepo (#2058)
Relates to #1951 

Currently our dev scripts (e.g. of `sit-onyx`) do not work in a clean
repository. The issue is that `sit-onyx` depends on other packages (like
icons) to be build first.

I aligned all `dev` scripts for all packages to be run with turborepo so
all dependencies are managed/build correctly.