chore: setup styelint for browser compatibility linting (#2128)
Relates to #1970 

First do `pnpm i`, then run `pnpm stylelint` to lint the Onyx CSS bundle
based on the supported browsers in .browserslistrc
This does NOT lint the styles in single Vue components, since the plugin
only supports CSS, no SCSS.
Currently only warnings are produced.

A sample output looks like this:
```
Unexpected browser feature "css-resize" is not supported by Safari on iOS
15.6-15.8,16.6-16.7,17.5,17.6-17.7,18.0,18.1
 ```

**How does this help ?** 

We get almost 200 warnings, mainly because of Firefox ESR. 

**Ideas:** 

We could run stylelint in the pipeline, parse the output and only fail on problems in certain situations. For example when only the latest Firefox or Safari version is affected.

Or we can just keep the command in package.json and just run it manually to validate.

---------

Co-authored-by: Oliver Müller <oliver.mueller2@mail.schwarz>
Co-authored-by: Lars Rickert <lars.rickert@mail.schwarz>