fix console warning for form elements about readonly target (#2555)
closes #2554

As described in #2554, the issue should already be fixed in Vue version
3.5.2. However, this does not seem to be the case. I assume this might
be because we are building the `sit-onyx` library (prod mode) and then
use it in an application (dev mode). Vue might handle prod and dev mode
differently.

As temporary fix for our onyx users, I renamed the template refs to be
different from the variable name.

I also created follow up tickets:

- #2556
- #2557

You can verify the fix by running the demo app and check the console.

## Checklist

- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)