fix(shared): change vite dependencies to optionalDependencies (#2565)
When installing the latest onyx version in a project that uses Vite 5,
it will show a peerDependency mismatch warning since `@sit-onyx/shared`
is declared to required Vite 6. However, the vite dependencies are only
needed when importing the shared vite config which is mainly done inside
our monorepo.

I changed the dependencies to be optionalDependencies to get rid of the
warning.

## Checklist

- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)