fix(OnyxFormElement): Fixed bug where the "optional" text didn't show up (#2708)
Relates to #2707 

- *fix(OnyxFormElement): Fixed bug where the "optional" text didn't show
up:*
The text (which is provided via an CSS ) custom property was not set
correctly in the plugin. The text needed to be enclosed in `"` to be
treated as valid text content.
- *fix(OnyxTextarea, OnyxInput): Fixed issue, where some supported props
were not displayed in API and Storybook:* Updated type definitions, some
props were accidentally removed.
- *feat(OnyxForm): `OnyxForm` now lets you also configure the
`requiredMarker` mode for the whole form:*
Updated form injected properties to also take care of `requiredMarker`,
so all these relevant form props are treated uniformly.
- Extended the `FormDemo` of the `demo-app` with some toggles, that let
you try out the form-level settings.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: JoCa96 <18483883+JoCa96@users.noreply.github.com>