feat(OnyxDataGrid): Implement cell renderer (#2679)
<!-- Is your PR related to an issue? Then please link it via the
"Relates to #" below. Else, remove it. -->

Relates to #2370 

- New 'type' property is added to the ColumnConfig type. With this
property a column can be of type 'string' or 'number'. Different styles
will be applied depending on the type of the column.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Jonathan Carle <jonathan_leo.carle@mail.schwarz>
Co-authored-by: JoCa96 <18483883+JoCa96@users.noreply.github.com>