feat(nuxt): re-enable cssnano calc after dependency was fixed (#2965)
<!-- Briefly describe the changes of this PR. -->
In the beginning we had to add a little workaround to the nuxt module as
one of the underlying dependencies for css bundling had a bug. This bug
was fixed here: https://github.com/postcss/postcss-calc/issues/210 and
the update by now made it into the current version of nuxt so we can
safely remove the workaround.

## Checklist

- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)