feat: implement OnyxNotificationCard component and notification center example (#2960)
Create a basic OnyxNotificationCard component that can be used to build
a notification center.
The notification center example will be extended in a follow up PR to
also use a OnyxNavBar with the OnyxNotificationDot.

Also made the default OnyxDrawer examples more generic to not confuse it
with the notification center

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
or functional test is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)

---------

Co-authored-by: Lars Rickert <lars.rickert@mail.schwarz>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: larsrickert <67898185+larsrickert@users.noreply.github.com>