feat(OnyxDataGrid): support column width definition (#3074)
- fix(OnyxDataGrid): fixed issue related to sticky columns and
restricted height
- it was necessary to use a `ResizeObserver` to track the height of the
table and pass it in via a custom property. Container query values
couldn't be used, as the table didn't render on Chromium when
container-type was used in combination with "display: grid".
- feat(OnyxDataGrid): support column width definition
  - was implemented as core funcitonality of the grid renderer
- the table is now rendered using CSS grid and subgrid. Subgrid is not
supported by older Chromium versions (<116), so the browserlists rules
have been loosened slightly. We don't need such a strict browserlists
anyways as we focus on business applications, where the browser are
defined on a corporate level
- feat(OnyxDataGrid/resizing): support double-click for auto-sizing

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: JoCa96 <18483883+JoCa96@users.noreply.github.com>