fix(OnyxDataGrid/resizing): fix error when using hiding columns with resizing feature (#3318)
**Cause:** The bug was caused by updating the `headers` map with a
`nullish` value when an element is unmounted or removed from the DOM.
When then trying to iterate over the entries of the map, the values were
expected to be dom elements.

**Fix:** When the `ref` function now receivs a nullish value, the entry
is isntead removed from the `header` map.

- Updated the `hideColumns` feature to not reorder readded columns. We
should only do this when we also offer an option to reorder columns.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: JoCa96 <18483883+JoCa96@users.noreply.github.com>