feat: release `@sit-onyx/flags` npm package for flag library (#3346)
Relates to #2701

Create a new npm package `@sit-onyx/flags` for our flag library. The
flags are imported from Figma (the same way as the icons).

I will update the docs to showcase all flags (similar to icons) in a
follow up PR.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)
- [x] All relevant changes are documented in the documentation app
(folder `apps/docs`) if needed

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>