fix(OnyxNavBar): fix runtime error when OnyxNavItems were mounted asynchronously (#3409)
Closes #3410 

**Cause:** Built-in `Teleport` is used to teleport the `OnyxNavItem`
content to the `OnyxMoreList` `options` slot (usually when there is not
enough horizontal space to display the item). So far the teleport
matched the target via an static id and the [`defer`
property](https://vuejs.org/guide/built-ins/teleport.html#deferred-teleport)
was used to ensure that the teleport is performed after everything is
rendered. Now, when `OnyxNavItems` were added async, the teleport was
tried and the `OnyxMoreList` didn't yet render the `options` slot the
`Teleport` component crashed with the error `can't access property
"type", instance is null`.

**Fix:** Instead of using a static id, we now make use of a TemplateRef.
This ref will only be defined when the was rendered and exists.
Additionally we disable the teleport as long as the ref is `undefined`