fix: prevent hydration errors for OnyxFlyout and OnyxTooltip (#3408)
closes #3398

Fix hydration errors in Nuxt when using the OnyxFlyout (or any component
that uses the OnyxFlyout internally like the OnyxUserMenu) or
OnyxTooltip components.

The issue is that we are currently immediately checking for the Anchor
API support which is executed on server side when using Nuxt so it will
always return false. I changed it to only check it on client side. Also
using `v-bind` in CSS caused hydration errors so I moved it to inline
styles instead.

Also fixed typo in `--onyx-flyout-max-width` CSS variable name which
caused the max width to not be applied

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)