feat(router): consider hash for active state (#3521)
Relates to #3450

Also consider hashes when determining the active state of a link when a
Vue Router is provided. E.g. the link `#some-headline` will now be
active when the current route is `/page-1#some-headline`

## Checklist

- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)