fix(OnyxDataGrid): do not prevent page scroll when sticky columns are active (#3639)
fixes #3637

Currently, the data grid / table blocks vertical page scroll when sticky
columns are active (see issue report #3637).
The behavior is fixed in this PR as well as providing a test to ensure
this behavior in the future.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)