fix(OnyxCheckbox): update checked and indeterminate icon color in dark mode (#3767)
Relates to #3650

Fix the checkmark and indeterminate icon color of the checkbox to update
in dark mode. Previously the color was always white which caused the
icons to not be visible in the kaufland dark theme.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
or functional test is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>