fix(nuxt): always provide onyx plugin (#3832)
Previously, the onyx plugin was only registered if Nuxt I18n was used.
This caused the issue that the router integration did not work when not
using i18n so we will now always provide the plugin but only pass the
i18n option if i18n exists.


## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)