feat: update onyx icon code replacement to use JavaScript instead of raw SVG imports (#3837)
<!-- Is your PR related to an issue? Then please link it via the
"Relates to #" below. Else, remove it. -->

closes #3836
relates to #1802

The usage of "getIconImportName" in our Storybook utils seemed to break
the deployed Storybook (see #1802). Weirdly, the utility works totally
fine when used with Vitest or in our documentation.

Regardlessly, we want to update the Storybook utils anyways in #3836 to
use the new JavaScript imports for the code snippets instead of raw SVG
imports. This was implemented in this PR so the "getIconImportName"
utility is no longer needed and therefore the deployed Storybook works
fine again.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)