fix(nuxt-docs): correctly pack files when publishing (#3847)
Relates to #3800 

After the release of #3800, I noticed that the Nuxt docs layer does not
work at all when used inside a project.
The reason is that due to the Nuxt 4 migration, all relevant files are
now placed inside the "app" directory but our "files" definition in
package.json did not include the directory when publishing so the file
were missing in the npm package.

## Checklist

- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)