fix(OnyxPageLayout): do not set grid max-width when noPadding is set (#3853)
Relates to #3665 

Due to a refactoring in version 1.0.0-beta.281, the OnyxPageLayout was
applying the max-width when the `noPadding` property was set.
This had the side effect that full-width content like e.g. hero images
could not be passed.

This behavior is fixed now so it works like prior version 1.0.0-beta.281

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
or functional test is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>