feat(nuxt)!: auto register onyx locales (#3866)
Relates to #3865

Remove the `registerOptions` property from the Nuxt module as it
involved manual user effort although it can be automated.
The module will now detect which locales the project is using/defining
and register the corresponding onyx locales accordingly.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)
- [x] All relevant changes are documented in the documentation app
(folder `apps/docs`) if needed

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>