fix(DataGridFeatures.useResizing): Fix issue where resizing didn't work with horizontal overflowing data grids (#3864)
- Closes #2746

Cause: The `showLastCol` ref was toggled in between clicks when trying
to double click. This caused a rerender, which invalidated the double
click.

Fix: We can simplfy the solution here and use a CSS only solution to
show the filler column. Now it is always being shown but can shrink to a
zero width, which makes it de-factor invisible. For a11y we hide the
filler column as it is only presentational.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: JoCa96 <18483883+JoCa96@users.noreply.github.com>