fix(@sit-onyx/vite-plugin-svg): Failing module resolution on Windows systems (#3940)
Relates to #3938

Cause: The typescript definition file was created with the name
`virtual:vite-plugin-svg.d.ts`.
But `:` is a [reserved character on
Windows](https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file#naming-conventions),
which causes issues with the module resolution on that OS.

Fix: We now use a file name without any special characters.