feat(OnyxTooltip)!: split up open and trigger property (#3953)
Relates to #

Previously, the `open` property of the OnyxTooltip and OnyxInfoTooltip
was used to both define the trigger type (hover, click) and set a
boolean for the open state.

This is changed now so:

- the `open` property is now just a boolean to control the open state.
Supports `v-model:open`.
- the new `trigger` property can be used to set the trigger type (hover,
click)


## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [ ] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
or functional test is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)