fix: implement workaround for boolean shorthands not working (#3957)
closes #2741

Previously, some boolean properties did not work when used as shorthand,
e.g. `<OnyxSelect multiple />` so they had to be explicitly set to
`true`.

We've implemented an internal workaround to fix this until the
[issue](https://github.com/SchwarzIT/onyx/issues/2741) is officially
fixed by the Vue core team.


## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)