refactor: replace demo app with Nuxt base setup (#4034)
Relates to #3969 

As mentioned in #3969 we will rebuild our demo app. This PR replaces the
current demo app which is kind of unused with a Nuxt application.

As of now, no page content exists but it will be added in follow up PRs
to keep things separated.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)