chore: fix demo app deployment (#4052)
Our demo app deployment is still failing. The issue was that the turbo
repo cache settings were not correct so the .output folder of the demo
app was never created because the cache was not rebuild.

I also defined the routes for all prod apps so we can configure them via
the manifest files.

- [Successful demo app deployment
run](https://github.com/SchwarzIT/onyx/actions/runs/17320797612)
- [Deployed demo app](https://demo.onyx.schwarz)