fix(OnyxDataGrid): do not check header checkbox when empty and no data is selected (#4122)
Relates to 4062

When using the `useSelection` feature of the OnyxDataGrid, the header
checkbox is now no longer checked when the data is empty and no rows are
checked (e.g. while loading async data from an API and showing skeleton
or when data was found at all).
Also, the checkbox is disabled when no data exists.

## Checklist

- [x] The added / edited code has been documented with
[JSDoc](https://jsdoc.app/about-getting-started)
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)