fix(OnyxDatepicker): Fix incorrect `v-model` type and behaviour which could cause runtime type mismatches (#4284)
Fix type mismatch between `update:modelValue` and `modelValue`, both are
now always using [ISO8601](https://en.wikipedia.org/wiki/ISO_8601)
strings.
This mismatch caused the value of a `ref` that is used with `v-model` to
be updated as string, even though it could have been typed as `Date` or
`number`.