fix(OnyxTab): define font-size and line-height for panel content (#4440)
We currently have some flaky screenshot tests for the OnyxCodeTabs. The
issue is propably that we use the `1lh` unit for the skeleton size but
the tab does not define any value for the line height.

This issue is fixed now by defining font size and line height for the
panel content inside the OnyxTab component.

## Checklist

- [x] If a new component is added, at least one [Playwright screenshot
test](https://github.com/SchwarzIT/onyx/actions/workflows/playwright-screenshots.yml)
or functional test is added
- [x] A changeset is added with `npx changeset add` if your changes
should be released as npm package (because they affect the library
usage)
- [x] I have performed a self review of my code ("Files changed" tab in
the pull request)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>