chore: fix eslint issues (#4441)
We currently have two [eslint
findings](https://github.com/SchwarzIT/onyx/security/code-scanning) for
some test files.
They are fixed in this PR and I also enabled a eslint rule to force
consistent usage of Vitest "test" function instead of mixing "test" and
"it". We are mostly already using "test" but there a a few files that
used it.